requirements.txt
langchain
langchain-community
numpy
pandas
ipykernel
transformers
pypdf
sentence_transformers
faiss-cpu
python-dotenv
openai
pdfplumber
langchain_huggingface
langchain-openai
